fix_3881_1 Modify test code for the BasicTestPerformance4Jdbc
